+++ packages/puppeteer-test-runner/src/App.tsx
@@ -1,32 +1,53 @@
 import React, {useEffect, useState} from 'react';
-import importConfig from "./importConfig";
 import fs from "fs"
 import path from "path"
-import {CreateTestProps} from "./createTest";
+import {CreateTestProps, PuppeteerConfig} from "./createTest";
+import puppeteer from "puppeteer";
+import createHelpers from "./createHelpers";
+
+
 function App() {
 
     const [files,setFiles] = useState<CreateTestProps[]>([])
+    const [puppeteerConfig,setPuppeteerConfig] = useState<PuppeteerConfig|null>(null)
 
     useEffect(() => {
         async function run(){
-            const conf = await importConfig()
-            const files = fs.readdirSync(conf.tests.dir)
-            // const a = "src\\__tests__\\test1"
-            // const l = path.resolve("C:\\Software\\femiadeniyi\\packages\\puppeteer-test-runner\\src\\__tests__\\test1")
-            // console.log(path.resolve(conf.tests.dir,files[0].replaceAll(".ts","")),"C:\\Software\\femiadeniyi\\packages\\puppeteer-test-runner\\src\\__tests__\\test1")
-            // let p = require(path.(`C:\\Software\\femiadeniyi\\packages\\puppeteer-test-runner\\${a}.ts`))
-            // import(p).then(console.log)
-            // setFiles(tests)
-            // console.log(tests)
+            console.log(window.puppeteer,"blo")
+            const browser = window.puppeteer.browser || await puppeteer.launch({
+                executablePath:"C:\\Program Files (x86)\\Microsoft\\Edge\\Application\\msedge.exe",
+                headless:false,
+                defaultViewport: null as unknown as undefined,
+                args:[
+                    '--inprivate',
+                    "--disable-web-security",
+                    "--allow-file-access-from-files",
+                    "--allow-file-access",
+                ]
+            })
+            const page = window.puppeteer.page || (await browser.pages())[0];
+            window.puppeteer.page = page
+            window.puppeteer.browser = browser
+            const helpers = await createHelpers({page})
+            setPuppeteerConfig({helpers,page,browser})
+            import("./__tests__/index.js").then(f => setFiles(f.default))
         }
         run()
     },[])
     return (
 
         <>
-            {files.map(f => {
+            {puppeteerConfig && files.map(f => {
 
-                return <button></button>
+                return (
+                    <button
+                        onClick={() => {
+                            f.test(puppeteerConfig)
+                        }}
+                    >
+                        {f.name}
+                    </button>
+                )
             })}
         </>
     );