+++ packages/puppeteer-test-runner/src/electron.ts
@@ -1,13 +1,4 @@
 import {app, BrowserWindow, dialog, Menu} from 'electron'
-import {importConfig} from "./importConfig";
-// import yargs from "yargs";
-//
-// const argv = yargs(process.argv.slice(2)).option('port', {
-//     string: true,
-//     demandOption: true
-// }).argv;
-//
-// console.log(argv)
 
 async function createWindow () {
     const win = new BrowserWindow({
@@ -18,9 +9,7 @@ async function createWindow () {
             nodeIntegration: true,
         },
     })
-    const config = await importConfig()
-
-    await win.loadURL(config.tests.url)
+    await win.loadURL(process.env.url!)
 }