+++ packages/firestore-api-helpers/CHANGELOG.md
@@ -5,7 +5,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.2.3](https://github.com/webdevshop/webdevshop-lerna/compare/v0.2.2...v0.2.3) (2021-02-08)
 
-**Note:** Version bump only for package firestore-api-helpers
+**Note:** Version bump only for package firestore-api-createHelpers
 
 
 
@@ -13,7 +13,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.2.2](https://github.com/webdevshop/webdevshop-lerna/compare/v0.2.1...v0.2.2) (2021-02-08)
 
-**Note:** Version bump only for package firestore-api-helpers
+**Note:** Version bump only for package firestore-api-createHelpers