+++ packages/puppeteer-test-helpers/CHANGELOG.md
@@ -5,7 +5,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.2.3](https://github.com/webdevshop/webdevshop-lerna/compare/v0.2.2...v0.2.3) (2021-02-08)
 
-**Note:** Version bump only for package puppeteer-test-helpers
+**Note:** Version bump only for package puppeteer-test-createHelpers
 
 
 
@@ -13,7 +13,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.2.2](https://github.com/webdevshop/webdevshop-lerna/compare/v0.2.1...v0.2.2) (2021-02-08)
 
-**Note:** Version bump only for package puppeteer-test-helpers
+**Note:** Version bump only for package puppeteer-test-createHelpers
 
 
 
@@ -21,7 +21,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.2.1](https://github.com/webdevshop/webdevshop-lerna/compare/v0.2.0...v0.2.1) (2021-02-08)
 
-**Note:** Version bump only for package @webdevshop/puppeteer-helpers
+**Note:** Version bump only for package @webdevshop/puppeteer-createHelpers