+++ packages/puppeteer-test-runner/CHANGELOG.md
@@ -35,4 +35,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 * **chore:** package.json ([c9e4f97](https://github.com/webdevshop/webdevshop/commit/c9e4f971b1b3de6b9f1e8b6bd97606fd6ba7775a))
 * **electron-test-runner:** init ([2a6ba27](https://github.com/webdevshop/webdevshop/commit/2a6ba2765dc853c89a686d2416bf9fdef4c7be12))
 * **puppeteer:** improving ([4ef8da9](https://github.com/webdevshop/webdevshop/commit/4ef8da922c4297c6cb3d6d4a2d93401729558cef))
-* **WDS-4:** publish to import helpers ([48c32c5](https://github.com/webdevshop/webdevshop/commit/48c32c516efc3f1a6c0d39df3d31f0ed95dd22e6))
+* **WDS-4:** publish to import createHelpers ([48c32c5](https://github.com/webdevshop/webdevshop/commit/48c32c516efc3f1a6c0d39df3d31f0ed95dd22e6))